flask-nats/requirements.txt
flask
redis
Flask-SSE
flask-blueprint
uwsgi
asyncio
asyncio-nats-client
